# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/drivers": {
      "get": {
        "operationId": "DriversController_findAll",
        "summary": "Liste des chauffeurs",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING_KYC",
                "PENDING_FIELD_VALIDATION",
                "APPROVED",
                "ACTIVE",
                "SUSPENDED",
                "AT_RISK",
                "TERMINATED",
                "BLACKLISTED"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "operationId": "DriversController_create",
        "summary": "Créer un profil chauffeur (à partir d'un User existant)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/drivers/{id}": {
      "get": {
        "operationId": "DriversController_findById",
        "summary": "Détail chauffeur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "patch": {
        "operationId": "DriversController_update",
        "summary": "Modifier les infos d'un chauffeur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "DriversController_remove",
        "summary": "Désactiver (soft delete) un chauffeur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/drivers/{id}/status": {
      "patch": {
        "operationId": "DriversController_updateStatus",
        "summary": "Changer le statut d'un chauffeur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/drivers/{id}/validate-kyc": {
      "post": {
        "operationId": "DriversController_validateKyc",
        "summary": "Valider le KYC d'un chauffeur (PENDING_KYC → PENDING_FIELD_VALIDATION)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/drivers/{id}/validate-field": {
      "post": {
        "operationId": "DriversController_validateField",
        "summary": "Valider la visite terrain d'un chauffeur (PENDING_FIELD_VALIDATION → APPROVED)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateFieldDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "drivers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Fleet Platform API",
    "description": "API de gestion de flotte VTC — Fleet Platform V2",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDriverDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID du User (compte téléphone) à lier au chauffeur"
          },
          "idCardNumber": {
            "type": "string"
          },
          "licenseNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "emergencyContact": {
            "type": "string"
          },
          "familyContacts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateDriverDto": {
        "type": "object",
        "properties": {
          "idCardNumber": {
            "type": "string"
          },
          "licenseNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "emergencyContact": {
            "type": "string"
          },
          "familyContacts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateDriverStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING_KYC",
              "PENDING_FIELD_VALIDATION",
              "APPROVED",
              "ACTIVE",
              "SUSPENDED",
              "AT_RISK",
              "TERMINATED",
              "BLACKLISTED"
            ],
            "description": "Nouveau statut du chauffeur"
          }
        },
        "required": [
          "status"
        ]
      },
      "ValidateFieldDto": {
        "type": "object",
        "properties": {
          "homeVisitDone": {
            "type": "boolean",
            "description": "La visite à domicile a été effectuée"
          },
          "homeVisitPhotoUrl": {
            "type": "string"
          },
          "gpsLocationLat": {
            "type": "number"
          },
          "gpsLocationLng": {
            "type": "number"
          },
          "environmentPhotoUrl": {
            "type": "string"
          },
          "resourcePersonName": {
            "type": "string"
          },
          "resourcePersonPhone": {
            "type": "string"
          },
          "familyContactsNotes": {
            "type": "string"
          },
          "managerComment": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}